+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 - ItemComboBoxBuilder.normalize() added.
 - Completion, CompletionDocument refactored to support accented character input in Linux.
 - ComboBoxBuilder and ItemComboBoxBuilder.normalize() parameter now boolean.
+- Completion.Builder added, Normalize removed, related refactoring.
 
 ## 0.18.21
 ### is.codion.common.db